+++ spec/hatchet/pipenv_spec.rb
@@ -24,12 +24,12 @@ RSpec.shared_examples 'aborts the build with a runtime not available message (Pi
       expect(clean_output(app.output)).to include(<<~OUTPUT)
         remote: -----> Python app detected
         remote: -----> Using Python version specified in Pipfile.lock
-        remote:  !     
-        remote:  !     Requested runtime 'python-#{requested_version}' is not available for this stack (#{app.stack}).
+        remote: 
+        remote:  !     Error: Requested runtime 'python-#{requested_version}' is not available for this stack (#{app.stack}).
         remote:  !     
         remote:  !     For a list of the supported Python versions, see:
         remote:  !     https://devcenter.heroku.com/articles/python-support#supported-runtimes
-        remote:  !     
+        remote: 
         remote:  !     Push rejected, failed to compile Python app.
       OUTPUT
     end
@@ -125,6 +125,8 @@ RSpec.describe 'Pipenv support' do
         expect(clean_output(app.output)).to match(Regexp.new(<<~OUTPUT))
           remote: -----> Python app detected
           remote: -----> Using Python version specified in Pipfile.lock
+          remote: 
+          remote:  !     Error: Python 3.6 is no longer supported.
           remote:  !     
           remote:  !     Python 3.6 reached upstream end-of-life on December 23rd, 2021, and is
           remote:  !     therefore no longer receiving security updates:
@@ -136,7 +138,7 @@ RSpec.describe 'Pipenv support' do
           remote:  !     
           remote:  !     For a list of the supported Python versions, see:
           remote:  !     https://devcenter.heroku.com/articles/python-support#supported-runtimes
-          remote:  !     
+          remote: 
           remote:  !     Push rejected, failed to compile Python app.
         OUTPUT
       end
@@ -151,6 +153,8 @@ RSpec.describe 'Pipenv support' do
         expect(clean_output(app.output)).to match(Regexp.new(<<~OUTPUT))
           remote: -----> Python app detected
           remote: -----> Using Python version specified in Pipfile.lock
+          remote: 
+          remote:  !     Error: Python 3.7 is no longer supported.
           remote:  !     
           remote:  !     Python 3.7 reached upstream end-of-life on June 27th, 2023, and is
           remote:  !     therefore no longer receiving security updates:
@@ -162,7 +166,7 @@ RSpec.describe 'Pipenv support' do
           remote:  !     
           remote:  !     For a list of the supported Python versions, see:
           remote:  !     https://devcenter.heroku.com/articles/python-support#supported-runtimes
-          remote:  !     
+          remote: 
           remote:  !     Push rejected, failed to compile Python app.
         OUTPUT
       end
@@ -273,12 +277,12 @@ RSpec.describe 'Pipenv support' do
         expect(clean_output(app.output)).to include(<<~OUTPUT)
           remote: -----> Python app detected
           remote: -----> Using Python version specified in Pipfile.lock
-          remote:  !     
-          remote:  !     Requested runtime '^3.12' is not available for this stack (#{app.stack}).
+          remote: 
+          remote:  !     Error: Requested runtime '^3.12' is not available for this stack (#{app.stack}).
           remote:  !     
           remote:  !     For a list of the supported Python versions, see:
           remote:  !     https://devcenter.heroku.com/articles/python-support#supported-runtimes
-          remote:  !     
+          remote: 
           remote:  !     Push rejected, failed to compile Python app.
         OUTPUT
       end
@@ -293,12 +297,12 @@ RSpec.describe 'Pipenv support' do
         expect(clean_output(app.output)).to include(<<~OUTPUT)
           remote: -----> Python app detected
           remote: -----> Using Python version specified in Pipfile.lock
-          remote:  !     
-          remote:  !     Requested runtime 'python-X.Y.Z' is not available for this stack (#{app.stack}).
+          remote: 
+          remote:  !     Error: Requested runtime 'python-X.Y.Z' is not available for this stack (#{app.stack}).
           remote:  !     
           remote:  !     For a list of the supported Python versions, see:
           remote:  !     https://devcenter.heroku.com/articles/python-support#supported-runtimes
-          remote:  !     
+          remote: 
           remote:  !     Push rejected, failed to compile Python app.
         OUTPUT
       end